diode-workstream/daffodil/requirements.txt
aws-cdk-lib==2.124.0
constructs>=10.0.0,<11.0.0
boto3==1.34.15
typer==0.15.2